refactor: extract giftPolaroidService
